Move count-words to apply/lib
