+++ controllers/apply/lib/count-words.js
@@ -6,15 +6,11 @@
  * If changing match this with word-count.vue
  * @param {string} text
  */
-function countWords(text) {
+module.exports = function countWords(text) {
     if (text) {
         const tokens = text.trim().match(/\S+/g) || [];
         return tokens.length;
     } else {
         return 0;
     }
-}
-
-module.exports = {
-    countWords
 };
